Simplify app bridge redirect
