--- app/assets/javascripts/shopify_app/app_bridge_redirect.js
@@ -1,10 +1,0 @@
-(function(window) {
-  function appBridgeRedirect(url) {
-    var normalizedLink = document.createElement('a');
-    normalizedLink.href = url;
-
-    open(normalizedLink.href, '_top');
-  }
-
-  window.appBridgeRedirect = appBridgeRedirect;
-})(window);
